+++ components/icon/component.tsx
@@ -11,7 +11,7 @@ interface IconProps {
 
 const SIZE = {
   sm: 'w-2\.5 h-2\.5',
-  md: 'w-3.5 h-auto',
+  md: 'w-3.5 h-3.5',
   lg: 'w-5 h-5',
   xlg: 'w-6 h-6',
 };
