+++ styles/index.css
@@ -6,7 +6,6 @@
   body {
     -webkit-font-smoothing: antialiased;
     font-family: 'Basier Circle';
-    font-weight: bold;
   }
 
   img,
